[2957] Clean up snippets in docs module for UI Components
