+++ docusaurus/docs/Android/03-ui/05-utility-components/04-attachment-gallery.mdx
@@ -80,5 +80,5 @@ activity?.activityResultRegistry?.let { registry ->
 destination.setData(attachmentGalleryItems = listOf(), attachmentIndex = 0)
 
 // Fire the navigation request
-ChatUI.instance().navigator.navigate(destination)
+ChatUI.navigator.navigate(destination)
 ```
