+++ stream-chat-android-docs/build.gradle
@@ -75,6 +75,7 @@ dependencies {
     implementation project(":stream-chat-android-ui-components")
     implementation project(":stream-chat-android-compose")
     implementation project(":stream-chat-android-pushprovider-firebase")
+    implementation project(":stream-chat-android-markdown-transformer")
 
     // Compose
     implementation Dependencies.composeCompiler
